+++ packages/pluggableWidgets/line-chart-native/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We have resolved an issue preventing Line chart from rendering correctly
+
 ## [3.0.1] - 2023-5-17
 
 ### Fixed